layout: <H2> navigation menu at right-hand margin
This menu contains links to all <H2> elements in the HTML document. It
degrades gracefully on lower screen resolutions (hidden when resolution
is less than or equal to 800x600; visible otherwise), works well in
text-only Web browsers such as Lynx and w3m, and also does not appear
when the document is printed or viewed in "Print Preview" mode.